Bootloader commands cleanup
