+++ boot/commands/commands.hpp
@@ -1,0 +1,18 @@
+#ifndef BOOT_COMMANDS_COMMANDS_HPP_
+#define BOOT_COMMANDS_COMMANDS_HPP_
+
+void Test();
+void TestSRAM();
+void TestEEPROM();
+
+void BootUpper();
+void BootToSelected();
+
+void SetBootIndex();
+
+void UploadApplication();
+void UploadSafeMode();
+
+void PrintBootTable();
+
+#endif /* BOOT_COMMANDS_COMMANDS_HPP_ */
